requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-bcrypt==1.0.1
flask-login==0.6.3
python-dotenv==1.0.1
pdfplumber==0.11.4
PyPDF2==3.0.1
spacy==3.7.6
transformers==4.45.2
torch==2.4.1
razorpay==1.4.2
tenacity==9.0.0
psycopg2-binary==2.9.9
requests==2.32.3
waitress==3.0.0